Refactored the connection form. Removed duplicate code
