+++ airbyte-webapp/src/components/CreateConnectionContent/index.tsx
@@ -1,0 +1,3 @@
+import CreateConnectionContent from "./CreateConnectionContent";
+
+export default CreateConnectionContent;
